requirements.txt
Flask
Flask-Cors
pyyaml